pyproject.toml
[project]
requires-python = ">=3.9"
dependencies = [
    "google-cloud-bigquery>=3.0.0",
    "apache-beam[gcp]>=2.50.0",
    "google-cloud-dlp>=3.0.0",
]
